Extract StorageUpdates
