+++ packages/liveblocks-core/src/__tests__/immutable.test.ts
@@ -7,7 +7,6 @@ import {
 } from "../immutable";
 import * as console from "../lib/fancy-console";
 import { LiveObject } from "../LiveObject";
-import type { StorageUpdate } from "../types";
 import type { BaseUserMeta } from "../types/BaseUserMeta";
 import type { ClientMsg } from "../types/ClientMsg";
 import { ClientMsgCode } from "../types/ClientMsg";
@@ -15,6 +14,7 @@ import type { Json, JsonObject } from "../types/Json";
 import type { LsonObject, ToJson } from "../types/Lson";
 import type { IdTuple, SerializedCrdt } from "../types/SerializedCrdt";
 import { ServerMsgCode } from "../types/ServerMsg";
+import type { StorageUpdate } from "../types/StorageUpdates";
 import {
   createSerializedList,
   createSerializedObject,
